+++ test/index.js
@@ -11,7 +11,7 @@ test('Set up', t => {
 test('Convert geojson to kml', t => {
   t.plan(1)
   const geojson = fs.createReadStream(`${__dirname}/fixtures/fc.geojson`)
-  const options = {format: 'kml', path: `${__dirname}/output`}
+  const options = {path: `${__dirname}/output`}
   const rows = []
   geojson
   .pipe(GeoXForm.createStream('kml', options))
@@ -24,7 +24,7 @@ test('Convert geojson to kml', t => {
 test('Convert geojson to csv', t => {
   t.plan(1)
   const geojson = fs.createReadStream(`${__dirname}/fixtures/fc.geojson`)
-  const options = {format: 'csv', path: `${__dirname}/output`}
+  const options = {path: `${__dirname}/output`}
   const rows = []
   geojson
   .pipe(GeoXForm.createStream('csv', options))
@@ -34,7 +34,25 @@ test('Convert geojson to csv', t => {
   .done(() => t.equal(rows.length, 101, 'CSV generated successfully'))
 })
 
-test('Teardown', t => {
-  Helper.after()
-  t.end()
+test('Convert geojson to shapefile', t => {
+  t.plan(1)
+  const geojson = fs.createReadStream(`${__dirname}/fixtures/fc.geojson`)
+  const options = {path: `${__dirname}/output`, name: 'test'}
+  const zipPath = `${__dirname}/output/xformed.zip`
+  geojson
+  .pipe(GeoXForm.createStream('zip', options))
+  .pipe(fs.createWriteStream(zipPath))
+  .on('finish', () => {
+    try {
+      fs.statSync(zipPath)
+      t.pass('Zip written successfully')
+    } catch (e) {
+      t.fail('Zip not written sucessfully')
+    }
+  })
 })
+
+// test('Teardown', t => {
+//   Helper.after()
+//   t.end()
+// })